Fix the controller possibly adopting foreign LBs on creation
We noticed two clusters with identical LBs in the same cloudscale organization managing the same Cloudscale LB.

As there is no way to atomically create a LoadBalancer in the Cloudscale API and then write back the created LBs UUID into the manifest, we need to check if we already created an LB but the manifest update failed in an earlier reconcile.
This lookup is done by the LB name generated from the controllers namespace and name. The second controller sees the LB and because namespace and name are equal thought it's an LB from an earlier reconcile.

This PR adds the Kubernetes resource UID as a tag to search for earlier unsuccessful reconciles.