Fix possibly adopting foreign LBs on creation (#26)
We noticed two clusters with identical LBs in the same cloudscale organization and zone managing the same Cloudscale LB.

As there is no way to atomically create a LoadBalancer in the Cloudscale API and then write back the created LBs UUID into the manifest, we need to check if we already created an LB but the manifest update failed in an earlier reconcile. This lookup is done by the LB name generated from the controllers namespace and name. The name/namespace combination is no longer unique if there are two clusters in the same cloudscale organization and zone. The second controller does not yet have a cloudscale UUID saved, looks up by name/namespace and adopts the other LB as the key matches.

This PR adds the Kubernetes resource UID as a tag to search for earlier unsuccessful reconciles.